Refactor: Classes
Converted the mostFreq file to a class, which is called in the main.cpp file.
This is so each new program I do I can call and run from the main file, rather than having a bunch of seperate mains ofc.
More OOP will be implemented later, with inheritance and such, but this is a basic convert.
Starting on a new prompt, nonrepeatChar, but need to convert it into a class before adding it back into project.