nonRepeat prompt complete and main change
completed the prompt for nonRepeat, though could still add a user input if i wanted to.

changed main so that the first case of the switch exists out, so i can keep adding the other options below it more simply.

still thinking about adding a parent class that the others can inherent from that could have user input and the basics included so that each one doesnt need iostream included and such.